orientations only sorta partially work.

+++ BlayneKinect3D/BlayneKinect3D/Blayne_Camera.h
@@ -51,5 +51,5 @@ private:
 	bool m_OnLeftEdge;
 	bool m_OnRightEdge;
 
-	glm::ivec2 m_mousePos;
+	glm::vec2 m_mousePos;
 };
